requirements.txt
asgiref==3.5.0
backports.zoneinfo==0.2.1
certifi==2021.10.8
cffi==1.15.0
cloudinary==1.29.0
cryptography==36.0.2
Django==4.0.3
django-environ==0.8.1
django-filter==21.1
djangorestframework==3.13.1
gunicorn==20.1.0
jwt==1.3.1
Pillow==9.0.1
pycparser==2.21
pytz==2021.3
six==1.16.0
sqlparse==0.4.2
urllib3==1.26.9
